// my-app/src/types/user.ts
export type User = {
  id: string
  name: string
  jobTitle: string
  location: string
  role: string | undefined
  email: string
  phone: string
  avatar: string
  ipAddress: string
  iban: string
  country: string
  countryCode: string
  language: string
  timeZone: string
  department: string
  postalCode: string
  companyName: string
  streetAddress: string
  username: string
  gender: string
  notes: string
}